Move serialize_signoff_requirements to a better place
