+++ src/auslib/web/admin/views/base.py
@@ -116,12 +116,3 @@ class AdminView(MethodView):
     def delete(self, *args, **kwargs):
         self.log.debug("processing DELETE request to %s" % request.path)
         return self._delete(*args, **kwargs)
-
-
-def serialize_signoff_requirements(requirements):
-    dct = {}
-    for rs in requirements:
-        signoffs_required = max(dct.get(rs["role"], 0), rs["signoffs_required"])
-        dct[rs["role"]] = signoffs_required
-
-    return dct
